\name{Estimation40k-package}
\alias{Estimation40k-package}
\alias{Estimation40k}
\docType{package}
\title{Simule les jets de dés de warhammer 40 000
\packageTitle{Estimation40k}
}
\description{
\packageDescription{Estimation40k}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Estimation40k}
\packageIndices{Estimation40k}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Estimation40k}

Maintainer: \packageMaintainer{Estimation40k}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
